Restructure the driver to support transactions better
Just about everything is rewritten to support transactions and do less
calculations and pin changes. Except with bitbang, where the
performance is on par, improvement is seen across the boards on SPI
(including AVR). Concurrent with the driver is an updated GFX library
to implement the optimized methods and give the highest score.
Both libs are 100% compatible and interchangeable with their original
variants.