Only include wiring_private.h if it exists
Some of the new Arduino cores do not include this file.  This includes the cores for the Arduino UNO Rev 4.

Two ways to fix this.
1) add it specifically to a list of cores to not include it, like this file has
#ifndef RASPI

Or use GCC __has_include to detect it and then include it.
With Arduino this will not work well for library headers, as the library search stuff will not see this as a dependency and as such won't add the -I to command line to find it.  But this file is in core which is always on the search path